requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
certifi==2021.5.30
charset-normalizer==2.0.2
Django==3.2.8
django-bootstrap-form==3.4
django-bootstrap4==3.0.1
django-crispy-forms==1.13.0
djangorestframework==3.12.4
idna==3.2
numpy==1.21.0
pandas==1.3.0
postgres==4.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
psycopg2-pool==1.1
python-dateutil==2.8.1
pytz==2021.1
readchar==3.0.4
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
urllib3==1.26.6
